#813 Add caching mechanism when fetching palette data
