+++ package/package.json
@@ -6,7 +6,7 @@
         "url": "https://github.com/railmapgen/rmg-palette.git",
         "directory": "package"
     },
-    "version": "1.0.1",
+    "version": "1.1.0",
     "description": "",
     "main": "./index.js",
     "scripts": {
