+++ package/src/cached-fetch.ts
@@ -1,0 +1,12 @@
+const RESPONSE_CACHE: Record<string, any> = {};
+
+export const cachedFetch = async (url: string, init?: RequestInit): Promise<any> => {
+    if (url in RESPONSE_CACHE) {
+        return RESPONSE_CACHE[url];
+    }
+
+    const res = await fetch(url, init);
+    const data = await res.json();
+    RESPONSE_CACHE[url] = data;
+    return data;
+};
